Add Venn diagrams visualizations; split stats.py
and reduce complexity of Statistics & Store
