+++ website/static/sass/page.sass
@@ -2,3 +2,15 @@
   margin-top: 10px
   font-size: 16px
   line-height: 1.63265
+
+.venn
+  text-align: center
+
+.venn-tooltip
+  position: absolute
+  text-align: center
+  background: #333
+  color: #eee
+  padding: 4px 20px
+  border-radius: 6px
+  opacity: 0
